+++ source/NORflash_OP.h
@@ -11,4 +11,4 @@ void IWRAM_CODE WriteFlash_with32word(u32 address,u8 *buffer,u32 size);
 u32 Loadfile2NOR(TCHAR *filename, u32 NORaddress,u16 have_patch,u8 SAVEMODE);
 void GenerateNORBitmap(u8 *bitmap);
 u32 GetFileListFromNor(void);
-int DeleteFile(int index);
+int DeleteFile(u8 index);
